requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-cloud-storage-transfer==1.9.2
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.6
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.1.0